Moved test predicate implementation into separate file
