+++ iris/src/test/scala/iris/dsl/RelationalOperatorsDSL.scala
@@ -8,32 +8,6 @@ import iris._
 class RelationalOperatorsDSL extends AnyFlatSpec {
   @Record case class User(username: String, age: Int)
 
-  object PredicateImpl extends PredicateDSL[String] {
-    def and(p1: String, p2: String): String = ???
-    def or(p1: String, p2: String): String = ???
-
-    def eq[T](fieldName: String, value: T): String =
-      s"$fieldName eq $value"
-
-    def neq[T](fieldName: String, value: T): String =
-      s"$fieldName neq $value"
-
-    def gt[T](fieldName: String, value: T): String =
-      s"$fieldName gt $value"
-
-    def gte[T](fieldName: String, value: T): String =
-      s"$fieldName gte $value"
-
-    def lt[T](fieldName: String, value: T): String =
-      s"$fieldName lt $value"
-
-    def lte[T](fieldName: String, value: T): String =
-      s"$fieldName lte $value"
-
-    def in[T](fieldName: String, value: Seq[T]): String =
-      s"$fieldName in [${value.mkString(", ")}]"
-  }
-
   def assertPredicate(p: Predicate[String], expected: String) =
     assert(p(PredicateImpl) == expected)
